Extract CLI stuff into other entrypoint
