+++ src/flashcard/flashcard.py
@@ -3,15 +3,11 @@
 """
 Flashcard game
 """
-from argparse import ArgumentParser
 import copy
-import logging
-import os
 import random
 
 from flashcard.cards import GuessStatus, Guess
 from flashcard.database import Db
-
 
 def display_flash_card(card):
     """Print info about card"""
@@ -73,53 +69,3 @@ def play(db_handle: "Db", deck, max_cards: int = 3, max_guesses: int = 5):
     correct_guesses = [g for g in guesses if g.status == GuessStatus.CORRECT]
     print(f"You correctly answered {len(correct_guesses)}/{all_card_count}.")
     db_handle.save_progress(guesses)
-
-
-def main(arguments):
-    """Main flashcard function"""
-    database_handle = Db("result.db")
-    database_handle.setup_database()
-
-    if args.deck.endswith(".json"):
-        if os.path.exists(args.deck):
-            database_handle.load_flash_cards(args.deck)
-    deck = database_handle.get_deck_from_database(args.deck)
-    display_deck_info(deck)
-
-    play(database_handle, deck, arguments.cards)
-
-
-if __name__ == "__main__":
-    logging.basicConfig(
-        level=logging.INFO,
-        filename="flashcard.log",
-        filemode="w",
-        format="[%(levelname)s] %(name)s -- %(message)s",
-    )
-    parser = ArgumentParser(description="Flashcard learning game")
-    parser.add_argument(
-        "deck", help="Path to deck of flashcards in json format or deck name"
-    )
-    parser.add_argument(
-        "-c",
-        "--cards",
-        help="Max number of flashcards",
-        type=int,
-        default=3,
-        required=False,
-    )
-    parser.add_argument(
-        "-t",
-        "--tries",
-        help="Max number of retries",
-        type=int,
-        default=3,
-        required=False,
-    )
-    args = parser.parse_args()
-    main(args)
-
-# USAGE:
-#   flashcard.py deck.json --cards 5 --tries 3
-#   flashcard.py deck.json -c 5 -t 3
-#   flashcard.py help
